feat(refacto): SJRA-1347 extract tab enums
