+++ frontend/apps/variant/src/entity/cases/slider/slider-uninterpreted-case-sheet.tsx
@@ -13,6 +13,7 @@ import SliderOccurrenceSubHeader from '@/components/base/slider/slider-occurrenc
 import SliderPatientRow from '@/components/base/slider/slider-patient-row';
 import SliderSheet from '@/components/base/slider/slider-sheet';
 import SliderSheetSkeleton from '@/components/base/slider/slider-sheet-skeleton';
+import { CaseEntityTabs } from '@/components/cores/types/case-tabs';
 import { useI18n } from '@/components/hooks/i18n';
 
 type CaseSliderSheetProps = {
@@ -93,7 +94,10 @@ function CaseSheetContent({ caseData, onPrevious, onNext, hasPrevious, hasNext }
           <Button
             variant="outline"
             onClick={() => {
-              window.open(`/case/entity/${caseResult.data?.case_id}?tab=variants&seq_id=${caseData.seq_id}`, '_blank');
+              window.open(
+                `/case/entity/${caseResult.data?.case_id}?tab=${CaseEntityTabs.Variants}&seq_id=${caseData.seq_id}`,
+                '_blank',
+              );
             }}
             size="sm"
           >
