+++ frontend/components/base/data-table/cells/germline-genome-frequency-cell.tsx
@@ -1,5 +1,6 @@
 import EmptyCell from '@/components/base/data-table/cells/empty-cell';
 import AnchorLink from '@/components/base/navigation/anchor-link';
+import { VariantEntityTabs } from '@/components/cores/types/variant-tabs';
 import { toExponentialNotation } from '@/components/lib/number-format';
 
 type GermlineGenomeFrequencyCellProps = {
@@ -16,7 +17,11 @@ function GermlineGenomeFrequencyCell({ pc, pf, locusId }: GermlineGenomeFrequenc
   const scientificNotationPC = toExponentialNotation(pc);
   const scientificNotationPF = toExponentialNotation(pf);
   return (
-    <AnchorLink size="sm" href={`/variants/entity/${locusId}?tab=patients&cases=InterpretedCases`} target="_blank">
+    <AnchorLink
+      size="sm"
+      href={`/variants/entity/${locusId}?tab=${VariantEntityTabs.Cases}&cases=InterpretedCases`}
+      target="_blank"
+    >
       {scientificNotationPC ? scientificNotationPC : pc} ({scientificNotationPF ? scientificNotationPF : pf})
     </AnchorLink>
   );
