+++ frontend/components/base/data-table/cells/participant-frequency-cell.tsx
@@ -1,5 +1,6 @@
 import EmptyCell from '@/components/base/data-table/cells/empty-cell';
 import AnchorLink from '@/components/base/navigation/anchor-link';
+import { VariantEntityTabs } from '@/components/cores/types/variant-tabs';
 import { toExponentialNotation } from '@/components/lib/number-format';
 
 type NumberCellProps = {
@@ -16,7 +17,7 @@ function ParticipantFrequencyCell({ value, locusId }: NumberCellProps) {
   const scientificNotation = toExponentialNotation(value);
 
   return (
-    <AnchorLink size="sm" href={`/variants/entity/${locusId}?tab=patients`} target="_blank">
+    <AnchorLink size="sm" href={`/variants/entity/${locusId}?tab=${VariantEntityTabs.Cases}`} target="_blank">
       {scientificNotation ? scientificNotation : value}
     </AnchorLink>
   );
