+++ frontend/components/base/data-table/cells/somatic-germline-frequency-cell.tsx
@@ -1,5 +1,6 @@
 import EmptyCell from '@/components/base/data-table/cells/empty-cell';
 import AnchorLink from '@/components/base/navigation/anchor-link';
+import { VariantEntityTabs } from '@/components/cores/types/variant-tabs';
 import { toExponentialNotation } from '@/components/lib/number-format';
 
 type NumberCellProps = {
@@ -15,7 +16,11 @@ function SomaticGermlineFrequencyCell({ pc, pf, locusId }: NumberCellProps) {
 
   const scientificNotationPF = toExponentialNotation(pf);
   return (
-    <AnchorLink size="sm" href={`/variants/entity/${locusId}?tab=patients&cases=InterpretedCases`} target="_blank">
+    <AnchorLink
+      size="sm"
+      href={`/variants/entity/${locusId}?tab=${VariantEntityTabs.Cases}&cases=InterpretedCases`}
+      target="_blank"
+    >
       {pc} ({scientificNotationPF ? scientificNotationPF : pf})
     </AnchorLink>
   );
