+++ frontend/components/base/occurrence/somatic-frequency-section.tsx
@@ -1,8 +1,10 @@
 import AnchorLink from '@/components/base/navigation/anchor-link';
+import { VariantEntityTabs } from '@/components/cores/types/variant-tabs';
 import { useI18n } from '@/components/hooks/i18n';
 
-import DetailSection, { DetailItem } from './detail-section';
 import EmptyField from '../information/empty-field';
+
+import DetailSection, { DetailItem } from './detail-section';
 
 export type SomaticFrequencySectionProps = {
   somatic_pc_tn_wgs?: number;
@@ -29,7 +31,11 @@ export default function SomaticFrequencySection({
         }
         value={
           somatic_pc_tn_wgs && somatic_pn_tn_wgs && somatic_pf_tn_wgs?.toExponential(2) ? (
-            <AnchorLink size="sm" href={`/variants/entity/${locusId}?tab=patients&cases=OtherCases`} target="_blank">
+            <AnchorLink
+              size="sm"
+              href={`/variants/entity/${locusId}?tab=${VariantEntityTabs.Cases}&cases=OtherCases`}
+              target="_blank"
+            >
               {`${somatic_pc_tn_wgs} / ${somatic_pn_tn_wgs} (${somatic_pf_tn_wgs?.toExponential(2)})`}
             </AnchorLink>
           ) : (
