+++ frontend/components/cores/types/case-tabs.ts
