+++ frontend/components/cores/types/variant-tabs.ts
